Breaking stuff so you know what to do.
